% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dec.R
\docType{data}
\name{dec}
\alias{dec}
\alias{dec_city_2010}
\alias{dec_city_2020}
\alias{dec_county_2010}
\alias{dec_county_2020}
\alias{dec_tract_2010}
\alias{dec_tract_2020}
\alias{dec_blockgroup_2010}
\alias{dec_blockgroup_2020}
\alias{dec_block_2010}
\alias{dec_block_2020}
\alias{dec_zcta_2010}
\alias{dec_zcta_2020}
\title{Decennial census population data}
\format{
\describe{
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{variable}{
\itemize{
\item Variable code
\item Data type: character
}
}
\item{value}{
\itemize{
\item Population count
\item Data type: double
}
}
\item{label}{
\itemize{
\item Descriptive variable label
\item Data type: character
}
}
\item{concept}{
\itemize{
\item Census table name
\item Data type: character
}
}
}

An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 4 rows and 6 columns.

An object of class \code{data.frame} with 4 rows and 6 columns.

An object of class \code{data.frame} with 168 rows and 6 columns.

An object of class \code{data.frame} with 185 rows and 6 columns.

An object of class \code{data.frame} with 451 rows and 6 columns.

An object of class \code{data.frame} with 510 rows and 6 columns.

An object of class \code{data.frame} with 12001 rows and 6 columns.

An object of class \code{data.frame} with 10730 rows and 6 columns.

An object of class \code{data.frame} with 62 rows and 6 columns.

An object of class \code{data.frame} with 59 rows and 6 columns.
}
\source{
US Census Bureau decennial census
}
\usage{
dec_city_2010

dec_city_2020

dec_county_2010

dec_county_2020

dec_tract_2010

dec_tract_2020

dec_blockgroup_2010

dec_blockgroup_2020

dec_block_2010

dec_block_2020

dec_zcta_2010

dec_zcta_2020
}
\description{
These data sets contain population data for Kansas City, Missouri, sourced
from the US Census Bureau's decennial census. Summary levels include city,
county, census tract, census block group, census block, and ZIP code
tabulation area (ZCTA).

Data sets were downloaded from \href{https://data.census.gov/}{data.census.gov} using \code{get_kc_pop()}.
See \href{https://www.census.gov/programs-surveys/decennial-census/guidance.html}{Guidance for Decennial Census of Population and Housing Data Users}
for information on how to use decennial census data.
}
\examples{
dec_city_2010

dec_tract_2020

}
\keyword{datasets}
