% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geoid2.R
\docType{data}
\name{geoid2}
\alias{geoid2}
\title{Geographic identifiers}
\format{
A list containing 8 dataframes.
\describe{
\item{state}{GEOID for Missouri}
\item{cbsa}{GEOID for Kansas City, KS-MO (core based statistical area)}
\item{place}{GEOID for Kansas City}
\item{county}{GEOIDs for counties intersecting with Kansas City}
\item{tract2010}{GEOIDs for census tracts intersecting with Kansas City
(2010)}
\item{tract2020}{GEOIDs for census tracts intersecting with Kansas City
(2020)}
\item{zcta2010}{GEOIDs for ZIP code tabulation areas intersecting with
Kansas City (2010)}
\item{zcta2020}{GEOIDs for ZIP code tabulation areas intersecting with
Kansas City (2020)}
}
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
geoid2
}
\description{
US Census Bureau geographic identifiers (GEOIDs) for areas that intersect
with Kansas City at several levels of geography.
}
\details{
Each element of \code{geoid2} is a dataframe containing GEOIDs for the areas
within a given geography level that intersect with Kansas City. Additional
details are included when available, such as area names, or useful, such as
county FIPS or census tract codes. For census tracts and ZCTAs, the
proportion of each area that overlaps with Kansas City is included.

See the US Census Bureau web page \href{https://www.census.gov/programs-surveys/geography/guidance/geo-identifiers.html}{Understanding Geographic Identifiers}
for more information on GEOIDs.
}
\keyword{datasets}
