% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geoids.R
\docType{data}
\name{geoids}
\alias{geoids}
\title{GEOIDs}
\format{
A list of GEOIDs for the following geographies
\describe{
\item{city}{Kansas City}
\item{county}{Cass, Clay, Jackson, and Platte Counties}
\item{tract}{Census tracts intersecting with Kansas City}
\item{blockgroup}{Census block groups intersecting with Kansas City}
\item{block}{Census blocks intersecting with Kansas City}
\item{zcta}{ZIP Code Tabulation Areas intersecting with Kansas City}
}
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
geoids
}
\description{
Geographic identifiers (GEOIDs) for geographic entities in kcData's spatial
data sets
}
\keyword{datasets}
