% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-pop.R
\name{get_kc_pop}
\alias{get_kc_pop}
\title{Get population data for Kansas City}
\usage{
get_kc_pop(
  dataset,
  geo = c("place", "county", "tract", "block group", "block", "zcta"),
  year,
  vars,
  var_match = c("fixed", "regex"),
  geoids = NULL,
  ...
)
}
\arguments{
\item{dataset}{The data set to download. Passed to the \code{dataset} argument in
\code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} and either \code{survey} in \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}
or \code{sumfile} in \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.}

\item{geo}{The data set geography (summary level).}

\item{year}{The data set year.}

\item{vars}{The variables to download. Either a vector of variable names or a
regular expression pattern.}

\item{var_match}{Use \code{"fixed"} to match variable names in \code{vars} exactly or
\code{"regex"} to match them to a regular expression pattern.}

\item{geoids}{A vector of GEOIDs to filter the output.}

\item{...}{Additional arguments passed to \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}} or
\code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}, such as \code{key} for the user's census API key.}
}
\value{
A dataframe.
}
\description{
Download population data for Kansas City from the decennial census or the
American Community Survey (ACS). Summary levels are place, county, census
tract, census block group, census block (decennial census only), and ZIP code
tabulation area (ZCTA). Data sets are downloaded from the US Census Bureau
using the \href{https://walker-data.com/tidycensus/index.html}{tidycensus} package.
}
\details{
ACS and census variables can be explored at the \href{https://censusreporter.org/}{Census Reporter} website
or by using \code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} to download variables for a
specific data set.

GEOIDs can be obtained from US Census Bureau shapefiles or from the data set
\code{geoid} in this package.

By default, \code{state = 29} is included in the call to the relevant tidycensus
function.
}
\section{Additional resources}{
\itemize{
\item \href{https://www.census.gov/programs-surveys/decennial-census.html}{Decennial census website}
\item \href{https://www.census.gov/programs-surveys/acs.html}{American Community Survey website}
\item \href{https://api.census.gov/data.html}{List of US Census Bureau data sets available via API}
\item \href{https://censusreporter.org/}{Census Reporter} (ACS tables and variables reference)
}
}

\examples{
\dontrun{
# Download place-level data from the 2024 1-year ACS
acs1_2024_place <- get_kc_pop(
  dataset = "acs1",
  geo = "place",
  year = 2024,
  vars = "^B01",
  var_match = "regex",
  geoids = geoid$place$geoid
)

# Download ZCTA-level data from the 2023 5-year ACS
acs5_2023_zcta <- get_kc_pop(
  dataset = "acs5",
  geo = "zcta",
  year = 2023,
  vars = "B01001_001",
  var_match = "fixed",
  geoids = geoid$zcta2020$geoid[geoid$zcta2020$overlap >= .1]
)

# Download census block-level data from the 2020 census
block <- get_kc_sf(
  geo = "block",
  year = 2024,
  intersect = "city",
  geometry = "full"
)

geoid_block <- block$GEOID20

census_2020_block <- get_kc_pop(
  dataset = "dhc",
  geo = "block",
  year = 2020,
  vars = "P12_001N",
  var_match = "fixed",
  geoids = geoid_block,
  county = geoid$county$county_fips
)
}

}
