% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-sf.R
\name{get_kc_sf}
\alias{get_kc_sf}
\title{Get simple features objects for Kansas City}
\usage{
get_kc_sf(
  geo = c("cbsa", "place", "county", "tract", "blockgroup", "block", "zcta"),
  year,
  intersect = c("city", "metro"),
  geometry = c("clipped", "full")
)
}
\arguments{
\item{geo}{The shapefile geography. \code{"place"} returns the city boundary and
\code{"cbsa"} returns the metro region boundary.}

\item{year}{The shapefile year.}

\item{intersect}{If \code{geo} is not \code{"place"} or \code{"cbsa"}, this is the larger
geography that determines the intersection for the geography selected in
\code{geo}. If \code{"metro"} is selected, geographies in both Kansas and Missouri will
be  returned.}

\item{geometry}{If \code{geo} is not \code{"place"} or \code{"cbsa"}, return either the full
geometries or geometries clipped to the geography indicated in \code{intersect}.}
}
\value{
A dataframe of class \code{sf} containing a simple feature list column.
}
\description{
Download a TIGER/Line shapefile containing
\itemize{
\item the Kansas City city boundary
\item the Kansas City metro region boundary
\item a set of smaller geographies intersecting with one of the above: county,
census tract, census block group, census block, or zip code tabulation area
(ZCTA)
}

Shapefiles are downloaded from the US Census Bureau using the \href{https://github.com/walkerke/tigris}{tigris}
package.
}
\details{
Geographies smaller than the city or metro region can contain the full
geometry or geometries clipped to the city or metro region boundary. For
these geographies, when \code{geometry = "clipped"} two additional variables are
added for the feature's full area and the area within the city or metro
region boundary.
}
\section{Additional resources}{
\itemize{
\item \href{https://www.census.gov/programs-surveys/geography/technical-documentation/complete-technical-documentation/tiger-geo-line.html}{TIGER/Line Shapefiles and TIGER/Line Files Technical Documentation}
\item \href{https://www2.census.gov/geo/pdfs/reference/geodiagram.pdf}{Standard Hierarchy of Census Geographic Entities}
\item \href{https://github.com/walkerke/tigris}{tigris package}
}
}

\examples{
\dontrun{
# Get the 2024 city boundary
city <- get_kc_sf(geo = "place", year = 2024)

# Get 2024 census tracts clipped to the city boundary
tracts <- get_kc_sf(geo = "tract", year = 2024)

# Get full 2024 ZCTAs that intersect with the KC metro region
zctas <- get_kc_sf(
  geo = "zcta",
  year = 2024,
  intersect = "metro",
  geometry = "full"
)
}

}
