% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-intercensal-estimates-2010-2019.R
\docType{data}
\name{intercensal_estimates_2010_2019}
\alias{intercensal_estimates_2010_2019}
\title{Intercensal population estimates, 2010-2019}
\format{
A dataframe with 10 rows and
2 columns.
\describe{
\item{year}{Year}
\item{estimate}{Resident population estimate as of July 1 of the year}
}
}
\source{
\href{https://www.census.gov/data/tables/time-series/demo/popest/intercensal-2010-2020-cities.html}{US Census Bureau, City and Town Intercensal Population Totals: 2010-2020}
}
\usage{
intercensal_estimates_2010_2019
}
\description{
Kansas City population estimates for 2010-2019. According to the \href{https://www.census.gov/programs-surveys/popest/technical-documentation/research/intercensal-estimates.html}{US Census Bureau}, "Once produced, the intercensal estimates become the preferred
series of data for the decade."
}
\keyword{datasets}
