% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-maps.R
\docType{data}
\name{maps}
\alias{maps}
\alias{map_city_2024}
\alias{map_tract_2024}
\alias{map_zcta_2024}
\title{Kansas City boundary maps}
\format{
TBD...

An object of class \code{sf} (inherits from \code{data.frame}) with 223 rows and 14 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 11 columns.
}
\source{
2024 TIGER/Line Shapefiles (US Census)
}
\usage{
map_city_2024

map_tract_2024

map_zcta_2024
}
\description{
Tables with map data for different geographies.
}
\keyword{datasets}
