% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-popest-v24.R
\docType{data}
\name{popest_v24}
\alias{popest_v24}
\title{Vintage 2024 population estimates}
\format{
\describe{
\item{name}{Area name}
\item{est_base_2020}{April 1, 2010, resident population estimates base}
\item{est_[year]}{Resident population estimate as of July 1 of the year}
}
}
\source{
\href{https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html}{US Census Bureau, City and Town Intercensal Population Totals: 2020-2024}
}
\usage{
popest_v24
}
\description{
Kansas City population estimates for 2021-2024. According to the \href{https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html}{US Census Bureau}, "Vintage 2024 is the most recent completed vintage and
consistent set of estimates."
}
\keyword{datasets}
