% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\docType{data}
\name{population}
\alias{population}
\alias{pop_city_2023acs5}
\alias{pop_tract_2023acs5}
\alias{pop_zcta_2023acs5}
\title{Kansas City population data sets}
\format{
TBD...

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 134915 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46585 rows and 7 columns.
}
\source{
2019-2023 American Community Survey 5-Year Estimates
}
\usage{
pop_city_2023acs5

pop_tract_2023acs5

pop_zcta_2023acs5
}
\description{
Tables with population data for different geographic levels.
}
\keyword{datasets}
