% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return-popest-table.R
\name{return_popest_table}
\alias{return_popest_table}
\title{Get population estimates}
\usage{
return_popest_table(years)
}
\arguments{
\item{years}{A numeric vector of years.}
}
\value{
A dataframe.
}
\description{
Create a table of population estimates using the intercensal population
estimates (\code{popest_ic}) and the most recent population estimate vintages
(\verb{popest_v<YY>}).
}
\examples{
return_popest_table(2015:2024)

}
