% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf-block.R
\docType{data}
\name{sf-block}
\alias{sf-block}
\alias{sf_block_2015}
\alias{sf_block_2016}
\alias{sf_block_2017}
\alias{sf_block_2018}
\alias{sf_block_2019}
\alias{sf_block_2020}
\alias{sf_block_2021}
\alias{sf_block_2022}
\alias{sf_block_2023}
\alias{sf_block_2024}
\title{Census block shapefiles}
\format{
Not all variables are present in all years. Some variable names end with the
2-digit census decade (e.g., "GEOID20").
\describe{
\item{STATEFP}{
\itemize{
\item State FIPS code
\item Data type: character
}
}
\item{COUNTYFP}{
\itemize{
\item County FIPS code
\item Data type: character
}
}
\item{TRACTCE}{
\itemize{
\item Census tract code
\item Data type: character
}
}
\item{BLOCKCE}{
\itemize{
\item Census tabulation block number
\item Data type: character
}
}
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{GEOIDFQ}{
\itemize{
\item Fully qualified GEOID
\item Data type: character
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{MTFCC}{
\itemize{
\item MAF/TIGER feature class code
\item Data type: character
}
}
\item{UR}{
\itemize{
\item Urban/rural indicator
\item Data type: character
}
}
\item{UACE}{
\itemize{
\item Urban area code
\item Data type: character
}
}
\item{UATYPE}{
\itemize{
\item Urban area type
\item Data type: character
}
}
\item{FUNCSTAT}{
\itemize{
\item Legal functional status
\item Data type: character
}
}
\item{ALAND}{
\itemize{
\item Land area
\item Data type: numeric
}
}
\item{AWATER}{
\itemize{
\item Water area
\item Data type: numeric
}
}
\item{INTPTLAT}{
\itemize{
\item Latitude of the internal point
\item Data type: character
}
}
\item{INTPTLON}{
\itemize{
\item Longitude of the internal point
\item Data type: character
}
}
\item{HOUSING}{
\itemize{
\item Housing unit count
\item Data type: character
}
}
\item{POP}{
\itemize{
\item Population count
\item Data type: character
}
}
\item{COUNTYFP}{
\itemize{
\item County FIPS code
\item Data type: character
}
}
\item{STATEFP}{
\itemize{
\item State FIPS code
\item Data type: character
}
}
\item{area}{
\itemize{
\item Feature area
\item Data type: double
}
}
\item{kc_area}{
\itemize{
\item Feature area within Kansas City boundary
\item Data type: double
}
}
\item{kc_area_pct}{
\itemize{
\item Percentage of feature area within Kansas City boundary
\item Data type: double
}
}
\item{geometry}{
\itemize{
\item Feature geometry
\item Data type: list
}
}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 12006 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 12009 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 12009 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 12009 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10730 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10730 rows and 19 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10732 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10755 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10756 rows and 21 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
sf_block_2015

sf_block_2016

sf_block_2017

sf_block_2018

sf_block_2019

sf_block_2020

sf_block_2021

sf_block_2022

sf_block_2023

sf_block_2024
}
\description{
TIGER/Line shapefiles for census blocks in Kansas City, Missouri, downloaded
from \href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} and modified using \code{\link[=get_kc_sf]{get_kc_sf()}}.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
