% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf-blockgroup.R
\docType{data}
\name{sf-blockgroup}
\alias{sf-blockgroup}
\alias{sf_blockgroup_2018}
\alias{sf_blockgroup_2019}
\alias{sf_blockgroup_2020}
\alias{sf_blockgroup_2021}
\alias{sf_blockgroup_2022}
\alias{sf_blockgroup_2023}
\alias{sf_blockgroup_2024}
\title{Census block group shapefiles}
\format{
TBD

An object of class \code{sf} (inherits from \code{data.frame}) with 451 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 510 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 510 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 510 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 510 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 510 rows and 17 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
sf_blockgroup_2018

sf_blockgroup_2019

sf_blockgroup_2020

sf_blockgroup_2021

sf_blockgroup_2022

sf_blockgroup_2023

sf_blockgroup_2024
}
\description{
TIGER/Line shapefiles for census block groups in Kansas City, Missouri,
downloaded from \href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} and modified using \code{\link[=get_kc_sf]{get_kc_sf()}}.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
