% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf-city.R
\docType{data}
\name{sf-city}
\alias{sf-city}
\alias{sf_city_2018}
\alias{sf_city_2019}
\alias{sf_city_2020}
\alias{sf_city_2021}
\alias{sf_city_2022}
\alias{sf_city_2023}
\alias{sf_city_2024}
\title{City boundary shapefiles}
\format{
\describe{
\item{STATEFP}{
\itemize{
\item State FIPS code
\item Data type: character
}
}
\item{PLACEFP}{
\itemize{
\item Place FIPS code
\item Data type: character
}
}
\item{PLACENS}{
\itemize{
\item Place NS code
\item Data type: character
}
}
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{GEOIDFQ}{
\itemize{
\item Fully qualified GEOID
\item Data type: character
\item 2023 table only
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{NAMELSAD}{
\itemize{
\item Geographic area name and legal/statistical area description
\item Data type: character
}
}
\item{LSAD}{
\itemize{
\item Legal/statistical area description
\item Data type: character
}
}
\item{CLASSFP}{
\itemize{
\item FIPS class code
\item Data type: character
}
}
\item{PCICBSA}{
\itemize{
\item Core based statistical area principal city indicator
\item Data type: character
}
}
\item{PCINECTA}{
\itemize{
\item New England city and town area principal city indicator
\item Data type: character
\item 2014-2022 tables only
}
}
\item{MTFCC}{
\itemize{
\item MAF/TIGER Feature Class Code
\item Data type: character
}
}
\item{FUNCSTAT}{
\itemize{
\item Legal functional status
\item Data type: character
}
}
\item{ALAND}{
\itemize{
\item Land area
\item Data type: numeric
}
}
\item{AWATER}{
\itemize{
\item Water area
\item Data type: numeric
}
}
\item{INTPTLAT}{
\itemize{
\item Latitude of the internal point
\item Data type: character
}
}
\item{INTPTLON}{
\itemize{
\item Longitude of the internal point
\item Data type: character
}
}
\item{geometry}{
\itemize{
\item Feature geometry
\item Data type: list
}
}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
sf_city_2018

sf_city_2019

sf_city_2020

sf_city_2021

sf_city_2022

sf_city_2023

sf_city_2024
}
\description{
TIGER/Line shapefiles for the city of Kansas City, Missouri, downloaded from
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} and modified using \code{\link[=get_kc_sf]{get_kc_sf()}}.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
