% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shp-city.R
\docType{data}
\name{shp-city}
\alias{shp-city}
\alias{shp_city_2014}
\alias{shp_city_2015}
\alias{shp_city_2016}
\alias{shp_city_2017}
\alias{shp_city_2018}
\alias{shp_city_2019}
\alias{shp_city_2020}
\alias{shp_city_2021}
\alias{shp_city_2022}
\alias{shp_city_2023}
\title{City boundary shapefiles}
\format{
\describe{
\item{STATEFP}{
\itemize{
\item State FIPS code
\item Data type: character
}
}
\item{PLACEFP}{
\itemize{
\item Place FIPS code
\item Data type: character
}
}
\item{PLACENS}{
\itemize{
\item Place NS code
\item Data type: character
}
}
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{GEOIDFQ}{
\itemize{
\item Fully qualified GEOID
\item Data type: character
\item 2023 table only
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{NAMELSAD}{
\itemize{
\item Geographic area name and legal/statistical area description
\item Data type: character
}
}
\item{LSAD}{
\itemize{
\item Legal/statistical area description
\item Data type: character
}
}
\item{CLASSFP}{
\itemize{
\item FIPS class code
\item Data type: character
}
}
\item{PCICBSA}{
\itemize{
\item Core based statistical area principal city indicator
\item Data type: character
}
}
\item{PCINECTA}{
\itemize{
\item New England city and town area principal city indicator
\item Data type: character
\item 2014-2022 tables only
}
}
\item{MTFCC}{
\itemize{
\item MAF/TIGER Feature Class Code
\item Data type: character
}
}
\item{FUNCSTAT}{
\itemize{
\item Legal functional status
\item Data type: character
}
}
\item{ALAND}{
\itemize{
\item Land area
\item Data type: numeric
}
}
\item{AWATER}{
\itemize{
\item Water area
\item Data type: numeric
}
}
\item{INTPTLAT}{
\itemize{
\item Latitude of the internal point
\item Data type: character
}
}
\item{INTPTLON}{
\itemize{
\item Longitude of the internal point
\item Data type: character
}
}
\item{geometry}{
\itemize{
\item TBD
\item Data type: character
}
}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles (2014-2023)
}
\usage{
shp_city_2014

shp_city_2015

shp_city_2016

shp_city_2017

shp_city_2018

shp_city_2019

shp_city_2020

shp_city_2021

shp_city_2022

shp_city_2023
}
\description{
TIGER/Line shapefiles for the city of Kansas City, Missouri, downloaded from
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} using the \href{https://github.com/walkerke/tigris}{tigris} package.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
