% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-shp-tract.R
\docType{data}
\name{shp-tract}
\alias{shp-tract}
\alias{shp_tract_2014}
\alias{shp_tract_2015}
\alias{shp_tract_2016}
\alias{shp_tract_2017}
\alias{shp_tract_2018}
\alias{shp_tract_2019}
\alias{shp_tract_2020}
\alias{shp_tract_2021}
\alias{shp_tract_2022}
\alias{shp_tract_2023}
\title{Census tract shapefiles}
\format{
\describe{
\item{STATEFP}{
\itemize{
\item State FIPS code
\item Data type: character
}
}
\item{COUNTYFP}{
\itemize{
\item County FIPS code
\item Data type: character
}
}
\item{TRACTCE}{
\itemize{
\item Census tract code
\item Data type: character
}
}
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{GEOIDFQ}{
\itemize{
\item Fully qualified GEOID
\item Data type: character
\item 2023 table only
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{NAMELSAD}{
\itemize{
\item Geographic area name and legal/statistical area description
\item Data type: character
}
}
\item{MTFCC}{
\itemize{
\item MAF/TIGER Feature Class Code
\item Data type: character
}
}
\item{FUNCSTAT}{
\itemize{
\item Legal functional status
\item Data type: character
}
}
\item{ALAND}{
\itemize{
\item Land area
\item Data type: numeric
}
}
\item{AWATER}{
\itemize{
\item Water area
\item Data type: numeric
}
}
\item{INTPTLAT}{
\itemize{
\item Latitude of the internal point
\item Data type: character
}
}
\item{INTPTLON}{
\itemize{
\item Longitude of the internal point
\item Data type: character
}
}
\item{geometry}{
\itemize{
\item TBD
\item Data type: character
}
}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 204 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 204 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 204 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 204 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 204 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 223 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 223 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 223 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 223 rows and 14 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles (2014-2023)
}
\usage{
shp_tract_2014

shp_tract_2015

shp_tract_2016

shp_tract_2017

shp_tract_2018

shp_tract_2019

shp_tract_2020

shp_tract_2021

shp_tract_2022

shp_tract_2023
}
\description{
TIGER/Line shapefiles for census tracts in Kansas City, Missouri, downloaded
from \href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} using the \href{https://github.com/walkerke/tigris}{tigris} package.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
