% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vintage-2024-estimates.R
\docType{data}
\name{vintage_2024_estimates}
\alias{vintage_2024_estimates}
\title{Vintage 2024 population estimates}
\format{
A dataframe with 5 rows and
2 columns.
\describe{
\item{year}{Year}
\item{estimate}{Resident population estimate as of July 1 of the year}
}
}
\source{
\href{https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html}{US Census Bureau, City and Town Intercensal Population Totals: 2020-2024}
}
\usage{
vintage_2024_estimates
}
\description{
Kansas City population estimates for 2021-2024. According to the \href{https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html}{US Census Bureau}, "Vintage 2024 is the most recent completed vintage and
consistent set of estimates."
}
\keyword{datasets}
