% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geoid2.R
\docType{data}
\name{geoid2}
\alias{geoid2}
\title{KC-area GEOIDs}
\format{
A list of dataframes.
\describe{
\item{state}{Missouri GEOID}
\item{cbsa}{Kansas City, KS-MO, GEOID}
\item{place}{Kansas City GEOID}
\item{county}{County GEOIDs and county FIPs codes}
\item{tract2011}{2011 census tract GEOIDs, tract codes, and proportion of
area overlapping with Kansas City}
\item{tract2020}{2020 census tract GEOIDs, tract codes, and proportion of
area overlapping with Kansas City}
\item{zcta2012}{2012 ZCTA GEOIDs and proportion of area overlapping with
Kansas City}
\item{zcta2020}{2020 ZCTA GEOIDs and proportion of area overlapping with
Kansas City}
}
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
geoid2
}
\description{
Geographic identifiers (GEOIDs) for the Kansas City area.
}
\details{
Each element of \code{geoid2} is a dataframe containing area names, if
applicable, and GEOIDs for a given geography level. \code{place}, \code{county},
\code{tract2011}, and \code{tract2020} contain the commonly used shorter codes (i.e.,
without the state FIPS code) for those levels. \code{tract2011}, \code{tract2020},
\code{zcta2012}, and \code{zcta2020} contain a variable \code{overlap} which is the
proportion of each area within the Kansas City boundary.

See the US Census Bureau web page \href{https://www.census.gov/programs-surveys/geography/guidance/geo-identifiers.html}{Understanding Geographic Identifiers}
for more information on GEOIDs.
}
\keyword{datasets}
