% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-pop.R
\name{get_kc_pop}
\alias{get_kc_pop}
\title{Get population data for Kansas City}
\usage{
get_kc_pop(
  dataset,
  geo = c("place", "county", "tract", "block group", "block", "zcta"),
  year,
  vars,
  var_match = c("fixed", "regex"),
  ...
)
}
\arguments{
\item{dataset}{The data set to download. Passed to the \code{dataset} argument in
\code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} and either \code{survey} in \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}
or \code{sumfile} in \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.}

\item{geo}{The geography of the data set to download. \code{"place"} returns the
city data.}

\item{year}{The year of the data set to download.}

\item{vars}{A vector of variable names or a regular expression to match
variable names for download.}

\item{var_match}{The type of matching to perform on variable names. \code{"fixed"}
matches the values in \code{vars} exactly, and \code{"regex"} matches to a regular
expression pattern in \code{vars}.}

\item{...}{Additional arguments passed to \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}} or
\code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.}
}
\value{
A dataframe.
}
\description{
Get population data for Kansas City
}
\examples{
\dontrun{
get_kc_pop(
  dataset = "acs5",
  geo = "place",
  year = 2023,
  vars = "^B01",
  var_match = "regex",
  show_call = TRUE
)
}

}
