% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-pop.R
\name{get_kc_pop}
\alias{get_kc_pop}
\title{Get population data for Kansas City}
\usage{
get_kc_pop(
  dataset,
  geo = c("place", "county", "tract", "block group", "block", "zcta"),
  year,
  vars,
  var_match = c("fixed", "regex"),
  ...
)
}
\arguments{
\item{dataset}{The data set to download. Passed to the \code{dataset} argument in
\code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} and either \code{survey} in \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}
or \code{sumfile} in \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.}

\item{geo}{The geography (summary level) of the data set to download.}

\item{year}{The year of the data set to download.}

\item{vars}{A vector of variable names or a regular expression to match
variable names for download.}

\item{var_match}{The type of matching to perform on variable names. \code{"fixed"}
matches the values in \code{vars} exactly, and \code{"regex"} matches to a regular
expression pattern in \code{vars}.}

\item{...}{Additional arguments passed to \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}} or
\code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.}
}
\value{
A dataframe.
}
\description{
Download population data for Kansas City from the decennial census or the
American Community Survey (ACS). Summary levels are place, county, census
tract, census block group, census block (decennial census only), and ZIP code
tabulation area (ZCTA). Data sets are downloaded from the US Census Bureau
using the \href{https://walker-data.com/tidycensus/index.html}{tidycensus} package.
}
\details{
Data sets for geographies other than "place" are filtered for geographic
units that intersect with Kansas City. These geographic units are determined
by the shapefile downloaded for the specified year. If no shapefile has been
downloaded for that year, an error is returned.
}
\section{Additional resources}{
\itemize{
\item \href{https://www.census.gov/programs-surveys/decennial-census.html}{Decennial census website}
\item \href{https://www.census.gov/programs-surveys/acs.html}{American Community Survey website}
\item \href{https://api.census.gov/data.html}{List of US Census Bureau data sets available via API}
\item \href{https://censusreporter.org/}{Census Reporter} (ACS tables and variables reference)
}
}

\examples{
\dontrun{
acs5_2023 <- get_kc_pop(
  dataset = "acs5",
  geo = "place",
  year = 2023,
  vars = "^B01",
  var_match = "regex"
)

dec_2020 <- get_kc_pop(
  dataset = "dhc",
  geo = "block",
  year = 2020,
  vars = "P12_001N",
  var_match = "fixed"
)
}

}
