% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-sf.R
\name{get_kc_sf}
\alias{get_kc_sf}
\title{Get simple features objects for Kansas City}
\usage{
get_kc_sf(
  year,
  geo = c("place", "county", "tract", "block group", "block", "zcta")
)
}
\arguments{
\item{year}{The year of the data set to download.}

\item{geo}{The geography of the data set to download. \code{"place"} returns the
city boundary geometry. All others return geometries that intersect with the
city boundary.}
}
\value{
A dataframe of class \code{sf} containing a simple feature list column.
}
\description{
Download a shapefile from the US Census Bureau for Kansas City or an
intersecting geography (counties, census tracts, census block groups, census
blocks, or zip code tabulation areas (ZCTAs)). Shapefiles are downloaded
using the \href{https://github.com/walkerke/tigris}{tigris} package.
}
\examples{
\dontrun{
get_kc_sf(2024, "tract")
}

}
