% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-sf.R
\name{get_kc_sf}
\alias{get_kc_sf}
\title{Get simple features objects for Kansas City}
\usage{
get_kc_sf(
  geo = c("place", "county", "tract", "block group", "block", "zcta"),
  year
)
}
\arguments{
\item{geo}{The geography of the data set to download. \code{"place"} returns the
city boundary geometry. All others return geometries that intersect with the
city boundary.}

\item{year}{The year of the data set to download.}
}
\value{
A dataframe of class \code{sf} containing a simple feature list column.
}
\description{
Download a TIGER/Line shapefile with boundaries for Kansas City or for a
geography intersecting with Kansas City: county, census tract, census block
group, census block, or zip code tabulation area (ZCTA). Files are downloaded
from the US Census Bureau using the \href{https://github.com/walkerke/tigris}{tigris} package.
}
\details{
Geographies that intersect with the city boundary are circumscribed to the
city boundary if they are not fully within the city.

These columns are added to the shapefiles for all geographies other than
"place":
\itemize{
\item \code{area}: the total area of the feature
\item \code{kc_area}: the area of the feature contained within the city boundary
\item \code{kc_area_pct}: the percentage of the feature within the city boundary
}

Area is calculated by \code{\link[sf:geos_measures]{sf::st_area()}}.
}
\section{Additional resources}{
\itemize{
\item \href{https://www.census.gov/programs-surveys/geography/technical-documentation/complete-technical-documentation/tiger-geo-line.html}{TIGER/Line Shapefiles and TIGER/Line Files Technical Documentation}
\item \href{https://www2.census.gov/geo/pdfs/reference/geodiagram.pdf}{Standard Hierarchy of Census Geographic Entities}
\item \href{https://github.com/walkerke/tigris}{tigris package}
}
}

\examples{
\dontrun{
get_kc_sf("tract", 2024)
}

}
