% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-kc-sf.R
\name{get_kc_sf}
\alias{get_kc_sf}
\title{Get simple features objects for Kansas City}
\usage{
get_kc_sf(
  geo = c("place", "county", "tract", "block group", "block", "zcta"),
  year
)
}
\arguments{
\item{geo}{The geography of the data set to download. \code{"place"} returns the
city boundary geometry. All others return geometries that intersect with the
city boundary.}

\item{year}{The year of the data set to download.}
}
\value{
A dataframe of class \code{sf} containing a simple feature list column.
}
\description{
Download a TIGER/Line shapefile from the US Census Bureau with boundaries for
Kansas City or for a geography intersecting with Kansas City (counties,
census tracts, census block groups, census blocks, or zip code tabulation
areas (ZCTAs)). Files are downloaded using the \href{https://github.com/walkerke/tigris}{tigris} package.
}
\details{
Geographies that intersect with the city boundary are circumscribed to the
city boundary if they are not fully within the city.

These columns are added to the shapefiles for all geographies other than
"place":
\itemize{
\item \code{area}: the total area of the feature
\item \code{kc_area}: the area of the feature contained within the city boundary
\item \code{kc_area_pct}: the percentage of the feature within the city boundary
}

Area is calculated by \code{\link[sf:geos_measures]{sf::st_area()}}.
}
\examples{
\dontrun{
get_kc_sf(2024, "tract")
}

}
