% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\docType{data}
\name{population}
\alias{population}
\alias{acs1_2014}
\alias{acs1_2015}
\alias{acs1_2016}
\alias{acs1_2017}
\alias{acs1_2018}
\alias{acs1_2019}
\alias{acs1_2021}
\alias{acs1_2022}
\alias{acs1_2023}
\alias{acs5_city_2014}
\alias{acs5_city_2015}
\alias{acs5_city_2016}
\alias{acs5_city_2017}
\alias{acs5_city_2018}
\alias{acs5_city_2019}
\alias{acs5_city_2020}
\alias{acs5_city_2021}
\alias{acs5_city_2022}
\alias{acs5_city_2023}
\alias{acs5_tract_2014}
\alias{acs5_tract_2015}
\alias{acs5_tract_2016}
\alias{acs5_tract_2017}
\alias{acs5_tract_2018}
\alias{acs5_tract_2019}
\alias{acs5_tract_2020}
\alias{acs5_tract_2021}
\alias{acs5_tract_2022}
\alias{acs5_tract_2023}
\alias{acs5_zcta_2014}
\alias{acs5_zcta_2015}
\alias{acs5_zcta_2016}
\alias{acs5_zcta_2017}
\alias{acs5_zcta_2018}
\alias{acs5_zcta_2019}
\alias{acs5_zcta_2020}
\alias{acs5_zcta_2021}
\alias{acs5_zcta_2022}
\alias{acs5_zcta_2023}
\title{Population data sets}
\format{
\describe{
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{variable}{
\itemize{
\item Variable ID consisting of the source table ID and variable code separated by an underscore
\item Data type: character
}
}
\item{estimate}{
\itemize{
\item Estimated population count
\item Data type: double
}
}
\item{moe}{
\itemize{
\item Margin of error for estimate
\item Data type: double
}
}
\item{label}{
\itemize{
\item Variable label consisting of the nested sequence of column labels from the source table
\item Data type: character
}
}
\item{concept}{
\itemize{
\item Source table name
\item Data type: character
}
}
\item{geography}{
\itemize{
\item Smallest geography for which a variable is summarized
\item Data type: character
\item 5-year ACS tables only
}
}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 850 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 821 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 529 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 667 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 647 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 107916 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140490 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140490 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148741 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144281 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36501 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48510 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48510 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51359 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49819 rows and 8 columns.
}
\source{
American Community Survey 1-Year and 5-Year data sets (2014-2023)
}
\usage{
acs1_2014

acs1_2015

acs1_2016

acs1_2017

acs1_2018

acs1_2019

acs1_2021

acs1_2022

acs1_2023

acs5_city_2014

acs5_city_2015

acs5_city_2016

acs5_city_2017

acs5_city_2018

acs5_city_2019

acs5_city_2020

acs5_city_2021

acs5_city_2022

acs5_city_2023

acs5_tract_2014

acs5_tract_2015

acs5_tract_2016

acs5_tract_2017

acs5_tract_2018

acs5_tract_2019

acs5_tract_2020

acs5_tract_2021

acs5_tract_2022

acs5_tract_2023

acs5_zcta_2014

acs5_zcta_2015

acs5_zcta_2016

acs5_zcta_2017

acs5_zcta_2018

acs5_zcta_2019

acs5_zcta_2020

acs5_zcta_2021

acs5_zcta_2022

acs5_zcta_2023
}
\description{
These data sets contain population data for Kansas City, Missouri, sourced
from the US Census Bureau's American Community Survey (ACS). In 1-year ACS
data sets, the summary level is Kansas City only. In 5-year ACS data sets,
available summary levels include city, census tract, and ZIP code tabulation
area (ZCTA).

Data sets were downloaded from \href{https://data.census.gov/}{data.census.gov} using the
\href{https://walker-data.com/tidycensus/index.html}{tidycensus} package. Tables... variables...

See \href{https://www.census.gov/programs-surveys/acs/guidance.html}{Guidance for Data Users} for information from census.gov on how to
use ACS data.
}
\examples{
acs1_2023

acs5_tract_2023

}
\keyword{datasets}
