% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-population.R
\docType{data}
\name{population}
\alias{population}
\alias{pop_2023_acs1}
\alias{pop_2023_acs5_city}
\alias{pop_2023_acs5_tract}
\alias{pop_2023_acs5_zcta}
\alias{pop_2022_acs1}
\alias{pop_2022_acs5_city}
\alias{pop_2022_acs5_tract}
\alias{pop_2022_acs5_zcta}
\alias{pop_2021_acs1}
\alias{pop_2021_acs5_city}
\alias{pop_2021_acs5_tract}
\alias{pop_2021_acs5_zcta}
\alias{pop_2020_acs1}
\alias{pop_2020_acs5_city}
\alias{pop_2020_acs5_tract}
\alias{pop_2020_acs5_zcta}
\alias{pop_2019_acs1}
\alias{pop_2019_acs5_city}
\alias{pop_2019_acs5_tract}
\alias{pop_2019_acs5_zcta}
\alias{pop_2018_acs1}
\alias{pop_2018_acs5_city}
\alias{pop_2018_acs5_tract}
\alias{pop_2018_acs5_zcta}
\alias{pop_2017_acs1}
\alias{pop_2017_acs5_city}
\alias{pop_2017_acs5_tract}
\alias{pop_2017_acs5_zcta}
\alias{pop_2016_acs1}
\alias{pop_2016_acs5_city}
\alias{pop_2016_acs5_tract}
\alias{pop_2016_acs5_zcta}
\alias{pop_2015_acs1}
\alias{pop_2015_acs5_city}
\alias{pop_2015_acs5_tract}
\alias{pop_2015_acs5_zcta}
\alias{pop_2014_acs1}
\alias{pop_2014_acs5_city}
\alias{pop_2014_acs5_tract}
\alias{pop_2014_acs5_zcta}
\title{Kansas City population data sets}
\format{
\describe{
\item{GEOID}{
\itemize{
\item The unique numeric identifier for a geography.
\item Data type: text
}
}
\item{NAME}{
\itemize{
\item The name of a geography.
\item Data type: text
}
}
\item{variable}{
\itemize{
\item The source table ID and variable code, separated by an underscore, of the estimate.
\item Data type: text
}
}
\item{estimate}{
\itemize{
\item The estimated population count.
\item Data type: integer
}
}
\item{moe}{
\itemize{
\item The margin of error for the estimate.
\item Data type: integer
}
}
\item{label}{
\itemize{
\item The nested sequence of column labels for the estimate from the source table.
\item Data type: text
}
}
\item{concept}{
\itemize{
\item The name of the source table containing the estimate.
\item Data type: text
}
}
\item{geography}{
\itemize{
\item The smallest geography for which a variable is summarized (5-year ACS only).
\item Data type: text
}
}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 647 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144281 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49819 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 850 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 667 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148741 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51359 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140490 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48510 rows and 8 columns.

An object of class \code{logical} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140490 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48510 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 128520 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43470 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 826 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 529 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 107916 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36501 rows and 8 columns.
}
\source{
American Community Survey 1-Year and 5-year data (2014-2023)
}
\usage{
pop_2023_acs1

pop_2023_acs5_city

pop_2023_acs5_tract

pop_2023_acs5_zcta

pop_2022_acs1

pop_2022_acs5_city

pop_2022_acs5_tract

pop_2022_acs5_zcta

pop_2021_acs1

pop_2021_acs5_city

pop_2021_acs5_tract

pop_2021_acs5_zcta

pop_2020_acs1

pop_2020_acs5_city

pop_2020_acs5_tract

pop_2020_acs5_zcta

pop_2019_acs1

pop_2019_acs5_city

pop_2019_acs5_tract

pop_2019_acs5_zcta

pop_2018_acs1

pop_2018_acs5_city

pop_2018_acs5_tract

pop_2018_acs5_zcta

pop_2017_acs1

pop_2017_acs5_city

pop_2017_acs5_tract

pop_2017_acs5_zcta

pop_2016_acs1

pop_2016_acs5_city

pop_2016_acs5_tract

pop_2016_acs5_zcta

pop_2015_acs1

pop_2015_acs5_city

pop_2015_acs5_tract

pop_2015_acs5_zcta

pop_2014_acs1

pop_2014_acs5_city

pop_2014_acs5_tract

pop_2014_acs5_zcta
}
\description{
These data sets contain population data for Kansas City, Missouri, sourced
from the US Census Bureau's American Community Survey (ACS). In 1-year ACS
data sets, the summary level is Kansas City only. In 5-year ACS data sets,
available summary levels include city, census tract, and ZIP code tabulation
area (ZCTA).

Data sets were obtained from \href{https://data.census.gov/}{data.census.gov} using the
\href{https://walker-data.com/tidycensus/index.html}{tidycensus} package.

See \href{https://www.census.gov/programs-surveys/acs/guidance.html}{Guidance for Data Users} for information from census.gov on how to
use ACS data.
}
\examples{
pop_2023_acs1

pop_2023_acs5_tract

}
\keyword{datasets}
