% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf-block.R
\docType{data}
\name{sf-block}
\alias{sf-block}
\alias{sf_block_2018}
\alias{sf_block_2019}
\alias{sf_block_2020}
\alias{sf_block_2021}
\alias{sf_block_2022}
\alias{sf_block_2023}
\alias{sf_block_2024}
\title{Census block shapefiles}
\format{
TBD

An object of class \code{sf} (inherits from \code{data.frame}) with 12009 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10730 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10730 rows and 19 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10732 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10755 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10756 rows and 21 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
sf_block_2018

sf_block_2019

sf_block_2020

sf_block_2021

sf_block_2022

sf_block_2023

sf_block_2024
}
\description{
TIGER/Line shapefiles for census blocks in Kansas City, Missouri, downloaded
from \href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} and modified using \code{\link[=get_kc_sf]{get_kc_sf()}}.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
