% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf-county.R
\docType{data}
\name{sf-county}
\alias{sf-county}
\alias{sf_county_2015}
\alias{sf_county_2016}
\alias{sf_county_2017}
\alias{sf_county_2018}
\alias{sf_county_2019}
\alias{sf_county_2020}
\alias{sf_county_2021}
\alias{sf_county_2022}
\alias{sf_county_2023}
\alias{sf_county_2024}
\title{County shapefiles}
\format{
TBD

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 22 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 22 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
sf_county_2015

sf_county_2016

sf_county_2017

sf_county_2018

sf_county_2019

sf_county_2020

sf_county_2021

sf_county_2022

sf_county_2023

sf_county_2024
}
\description{
TIGER/Line shapefiles for counties in Kansas City, Missouri, downloaded from
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} and modified using \code{\link[=get_kc_sf]{get_kc_sf()}}.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
