% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf-tract.R
\docType{data}
\name{sf-tract}
\alias{sf-tract}
\alias{sf_tract_2018}
\alias{sf_tract_2019}
\alias{sf_tract_2020}
\alias{sf_tract_2021}
\alias{sf_tract_2022}
\alias{sf_tract_2023}
\alias{sf_tract_2024}
\title{Census tract shapefiles}
\format{
\describe{
\item{STATEFP}{
\itemize{
\item State FIPS code
\item Data type: character
}
}
\item{COUNTYFP}{
\itemize{
\item County FIPS code
\item Data type: character
}
}
\item{TRACTCE}{
\itemize{
\item Census tract code
\item Data type: character
}
}
\item{GEOID}{
\itemize{
\item Geographic identifier
\item Data type: character
}
}
\item{GEOIDFQ}{
\itemize{
\item Fully qualified GEOID
\item Data type: character
\item 2023 table only
}
}
\item{NAME}{
\itemize{
\item Geographic area name
\item Data type: character
}
}
\item{NAMELSAD}{
\itemize{
\item Geographic area name and legal/statistical area description
\item Data type: character
}
}
\item{MTFCC}{
\itemize{
\item MAF/TIGER Feature Class Code
\item Data type: character
}
}
\item{FUNCSTAT}{
\itemize{
\item Legal functional status
\item Data type: character
}
}
\item{ALAND}{
\itemize{
\item Land area
\item Data type: numeric
}
}
\item{AWATER}{
\itemize{
\item Water area
\item Data type: numeric
}
}
\item{INTPTLAT}{
\itemize{
\item Latitude of the internal point
\item Data type: character
}
}
\item{INTPTLON}{
\itemize{
\item Longitude of the internal point
\item Data type: character
}
}
\item{geometry}{
\itemize{
\item Feature geometry
\item Data type: list
}
}
\item{area}{
\itemize{
\item Feature area
\item Data type: double
}
}
\item{kc_area}{
\itemize{
\item Feature area within Kansas City boundary
\item Data type: double
}
}
\item{kc_area_pct}{
\itemize{
\item Percentage of feature area within Kansas City boundary
\item Data type: double
}
}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 168 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 185 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 185 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 185 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 185 rows and 17 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 185 rows and 17 columns.
}
\source{
US Census Bureau TIGER/Line Shapefiles (2018-2024)
}
\usage{
sf_tract_2018

sf_tract_2019

sf_tract_2020

sf_tract_2021

sf_tract_2022

sf_tract_2023

sf_tract_2024
}
\description{
TIGER/Line shapefiles for census tracts in Kansas City, Missouri, downloaded
from \href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{census.gov} and modified using \code{\link[=get_kc_sf]{get_kc_sf()}}.

The US Census Bureau uses the Global Coordinate System North American Datum
of 1983 (GCS NAD83/EPSG: 4269).
}
\keyword{datasets}
